1 less api call

+++ spec/spec_helper.rb
@@ -48,8 +48,7 @@ RSpec.configure do |config|
       begin
         print "."
         ids = Cloudinary::Api.resources_by_tag('test_env', max_results: 100)["resources"].map{|r| r["public_id"] }
-        Cloudinary::Api.delete_resources(ids)
-        sleep 1
+        Cloudinary::Api.delete_resources(ids) unless ids.empty?
       end until ids.empty?
       print " (done)"
     rescue Cloudinary::Api::RateLimited => e
